requirements/base.txt
Django==1.11.3
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-extensions==1.7.9
django-filter==1.0.4
django-rest-swagger==2.1.1
djangorestframework==3.6.3
djangorestframework-jsonapi==2.1.1
google-api-python-client==1.5.1
psycopg2-binary
httplib2==0.9.2
mongoengine==0.10.6
oauth2client==2.2.0
oauthlib==1.0.3
PyJWT==1.4.0
pymongo==3.3.0
python-openid==2.2.5
social-auth-app-django==1.2.0
pytz==2015.6
requests==2.11.1
requests-oauthlib==0.5.0
six==1.10.0
cryptography==2.1.3
fabric==1.14.0
celery==4.1.1
django-celery-results==1.0.1
redis==2.10.5
